Added Simple Camera Switching In-Game
Now In Game, going up the stairs of the structure will now switch the camera to a different angle, same said for going all the way to the end of the structure on top. Haven't Added a way  to reset the cameras upon falling down to the lower ground. Tried Adding a working Resolution Option to the Settings menu however did not succeed